requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
fastapi[standard]==0.115.0